fix: [CS 5.18]: recover pending deposits
When a deposit is made and a strategy is removed before that deposit is flushed,
the deposit is simply redistributed among other strategies. This leaves one edge
case, removing the last strategy of a smart vault. For this case, an emergency
recovery is implemented that sends such pending deposits to emergency wallet.
Additionally, flush should be disabled for smart vaults with no valid strategy.

- can no longer flush or redeem on ghost vaults
- added emergency method to recover pending deposits
- added tests for what happens with assets when removing a strategy in
  various states of deposit / withdrawal